Moved beans around
