+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationPageServiceImpl.java
@@ -4,6 +4,7 @@ import com.appsmith.server.acl.PolicyGenerator;
 import com.appsmith.server.actioncollections.base.ActionCollectionService;
 import com.appsmith.server.helpers.GitFileUtils;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.repositories.ActionCollectionRepository;
@@ -33,6 +34,7 @@ public class ApplicationPageServiceImpl extends ApplicationPageServiceCEImpl imp
             SessionUserService sessionUserService,
             WorkspaceRepository workspaceRepository,
             LayoutActionService layoutActionService,
+            UpdateLayoutService updateLayoutService,
             AnalyticsService analyticsService,
             PolicyGenerator policyGenerator,
             ApplicationRepository applicationRepository,
@@ -53,13 +55,13 @@ public class ApplicationPageServiceImpl extends ApplicationPageServiceCEImpl imp
             NewPageRepository newPageRepository,
             DatasourceRepository datasourceRepository,
             DatasourcePermission datasourcePermission) {
-
         super(
                 workspaceService,
                 applicationService,
                 sessionUserService,
                 workspaceRepository,
                 layoutActionService,
+                updateLayoutService,
                 analyticsService,
                 policyGenerator,
                 applicationRepository,
