+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/LayoutCollectionServiceImpl.java
@@ -2,6 +2,7 @@ package com.appsmith.server.services;
 
 import com.appsmith.server.actioncollections.base.ActionCollectionService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.refactors.applications.RefactoringSolution;
@@ -19,6 +20,7 @@ public class LayoutCollectionServiceImpl extends LayoutCollectionServiceCEImpl i
     public LayoutCollectionServiceImpl(
             NewPageService newPageService,
             LayoutActionService layoutActionService,
+            UpdateLayoutService updateLayoutService,
             RefactoringSolution refactoringSolution,
             ActionCollectionService actionCollectionService,
             NewActionService newActionService,
@@ -30,6 +32,7 @@ public class LayoutCollectionServiceImpl extends LayoutCollectionServiceCEImpl i
         super(
                 newPageService,
                 layoutActionService,
+                updateLayoutService,
                 refactoringSolution,
                 actionCollectionService,
                 newActionService,
