+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/LayoutActionServiceCE.java
@@ -2,22 +2,10 @@ package com.appsmith.server.services.ce;
 
 import com.appsmith.external.helpers.AppsmithEventContext;
 import com.appsmith.external.models.ActionDTO;
-import com.appsmith.server.domains.Layout;
 import com.appsmith.server.dtos.ActionMoveDTO;
-import com.appsmith.server.dtos.LayoutDTO;
-import com.appsmith.server.dtos.UpdateMultiplePageLayoutDTO;
-import net.minidev.json.JSONObject;
 import reactor.core.publisher.Mono;
 
 public interface LayoutActionServiceCE {
-
-    Mono<LayoutDTO> updateLayout(String pageId, String applicationId, String layoutId, Layout layout);
-
-    Mono<LayoutDTO> updateLayout(
-            String defaultPageId, String defaultApplicationId, String layoutId, Layout layout, String branchName);
-
-    Mono<Integer> updateMultipleLayouts(
-            String defaultApplicationId, String branchName, UpdateMultiplePageLayoutDTO updateMultiplePageLayoutDTO);
 
     Mono<ActionDTO> moveAction(ActionMoveDTO actionMoveDTO);
 
@@ -34,8 +22,6 @@ public interface LayoutActionServiceCE {
     Mono<ActionDTO> setExecuteOnLoad(String id, Boolean isExecuteOnLoad);
 
     Mono<ActionDTO> setExecuteOnLoad(String defaultActionId, String branchName, Boolean isExecuteOnLoad);
-
-    JSONObject unescapeMongoSpecialCharacters(Layout layout);
 
     Mono<ActionDTO> createAction(ActionDTO action);
