+++ app/server/appsmith-server/src/test/java/com/appsmith/server/refactors/RefactoringSolutionTest.java
@@ -15,6 +15,7 @@ import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.PluginExecutorHelper;
 import com.appsmith.server.imports.internal.ImportApplicationService;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.refactors.applications.RefactoringSolution;
@@ -79,6 +80,9 @@ public class RefactoringSolutionTest {
 
     @Autowired
     LayoutActionService layoutActionService;
+
+    @Autowired
+    UpdateLayoutService updateLayoutService;
 
     @Autowired
     RefactoringSolution refactoringSolution;
@@ -175,7 +179,7 @@ public class RefactoringSolutionTest {
 
         layout.setDsl(dsl);
         layout.setPublishedDsl(dsl);
-        layoutActionService
+        updateLayoutService
                 .updateLayout(pageId, testApp.getId(), layout.getId(), layout)
                 .block();
